requirements.txt
requests
colorama
pyautogui
pynput
Pillow